Refactor EmployeeProvider to use utility helpers
- Extract employee data loading logic into loadEmployees in utils.
- Extract color application logic into applyColors in utils.
- Simplify useEffect by composing stored employees, color application, and week permutation.
- Improve readability and reusability of employee management code.
